OpenToonz Patches thru 2/9/2023

+++ toonz/sources/common/tcolor/tcolorfunctions.cpp
@@ -28,10 +28,10 @@ TGenericColorFunction::TGenericColorFunction(const double m[4],
 //---------------------------------------
 
 TPixel32 TGenericColorFunction::operator()(const TPixel32 &color) const {
-  return TPixel32(tcrop(m_m[0] * color.r + m_c[0], 0.0, 255.0),
-                  tcrop(m_m[1] * color.g + m_c[1], 0.0, 255.0),
-                  tcrop(m_m[2] * color.b + m_c[2], 0.0, 255.0),
-                  tcrop(m_m[3] * color.m + m_c[3], 0.0, 255.0));
+  return TPixel32((int)tcrop(m_m[0] * color.r + m_c[0], 0.0, 255.0),
+                  (int)tcrop(m_m[1] * color.g + m_c[1], 0.0, 255.0),
+                  (int)tcrop(m_m[2] * color.b + m_c[2], 0.0, 255.0),
+                  (int)tcrop(m_m[3] * color.m + m_c[3], 0.0, 255.0));
 }
 
 //---------------------------------------
@@ -46,7 +46,7 @@ bool TGenericColorFunction::getParameters(Parameters &p) const {
 //---------------------------------------
 
 TPixel32 TTranspFader::operator()(const TPixel32 &color) const {
-  return TPixel32(color.r, color.g, color.b, m_transp * color.m);
+  return TPixel32(color.r, color.g, color.b, (int)(m_transp * (double)color.m));
 }
 
 //---------------------------------------
